Refactor cloud_api_client for session based authentication
