+++ crates/cloud_api_client/src/cloud_api_client.rs
@@ -2,8 +2,6 @@ mod llm_token;
 mod websocket;
 
 use std::sync::Arc;
-#[cfg(target_family = "wasm")]
-use std::time::Duration;
 
 use anyhow::{Result, anyhow};
 pub use cloud_api_types::*;
@@ -18,12 +16,15 @@ use thiserror::Error;
 
 pub use llm_token::LlmApiToken;
 
-#[cfg(target_family = "wasm")]
-const WEBSOCKET_CONNECT_TIMEOUT: Duration = Duration::from_secs(30);
-
 struct Credentials {
     user_id: u32,
     access_token: String,
+}
+
+#[derive(Clone, Copy)]
+enum Authentication {
+    Credentials,
+    Session,
 }
 
 #[derive(Debug, Error)]
@@ -61,6 +62,7 @@ pub enum ClientApiError {
 pub struct CloudApiClient {
     credentials: RwLock<Option<Credentials>>,
     http_client: Arc<HttpClientWithUrl>,
+    authentication: Authentication,
 }
 
 impl CloudApiClient {
@@ -68,6 +70,15 @@ impl CloudApiClient {
         Self {
             credentials: RwLock::new(None),
             http_client,
+            authentication: Authentication::Credentials,
+        }
+    }
+
+    pub fn new_with_session_authentication(http_client: Arc<HttpClientWithUrl>) -> Self {
+        Self {
+            credentials: RwLock::new(None),
+            http_client,
+            authentication: Authentication::Session,
         }
     }
 
@@ -112,80 +123,6 @@ impl CloudApiClient {
 
         self.send_authenticated_json_request(request_builder, AsyncBody::default())
             .await
-    }
-
-    #[cfg(not(target_family = "wasm"))]
-    pub fn connect(
-        self: &Arc<Self>,
-        cx: &gpui::App,
-    ) -> Result<gpui::Task<Result<crate::websocket::Connection>>> {
-        use anyhow::Context as _;
-        use cloud_api_types::websocket_protocol::{PROTOCOL_VERSION, PROTOCOL_VERSION_HEADER_NAME};
-
-        let mut connect_url = self
-            .http_client
-            .build_zed_cloud_url("/client/users/connect")?;
-        connect_url
-            .set_scheme(match connect_url.scheme() {
-                "https" => "wss",
-                "http" => "ws",
-                scheme => Err(anyhow!("invalid URL scheme: {scheme}"))?,
-            })
-            .map_err(|_| anyhow!("failed to set URL scheme"))?;
-
-        let credentials = self.credentials.read();
-        let credentials = credentials.as_ref().context("no credentials provided")?;
-        let authorization_header = format!("{} {}", credentials.user_id, credentials.access_token);
-
-        Ok(gpui_tokio::Tokio::spawn_result(cx, async move {
-            let websocket = yawc::WebSocket::connect(connect_url)
-                .with_request(
-                    request::Builder::new()
-                        .header("Authorization", authorization_header)
-                        .header(PROTOCOL_VERSION_HEADER_NAME, PROTOCOL_VERSION.to_string()),
-                )
-                .await?;
-
-            Ok(websocket::Connection::new(websocket))
-        }))
-    }
-
-    #[cfg(target_family = "wasm")]
-    pub fn connect(
-        self: &Arc<Self>,
-        cx: &gpui::App,
-    ) -> Result<gpui::Task<Result<crate::websocket::Connection>>> {
-        use cloud_api_types::websocket_protocol::PROTOCOL_VERSION;
-        use futures::FutureExt as _;
-
-        let client = self.clone();
-        let executor = cx.background_executor().clone();
-        Ok(cx.spawn(async move |_cx| {
-                let request_builder = Request::builder().method(Method::POST).uri(
-                    client
-                        .http_client
-                        .build_zed_cloud_url("/client/users/websocket_connection")?
-                        .as_ref(),
-                );
-                let CreateWebSocketConnectionResponse { url } = client
-                    .send_authenticated_json_request(
-                        request_builder,
-                        Json(CreateWebSocketConnectionBody {
-                            protocol_version: PROTOCOL_VERSION,
-                        }),
-                    )
-                    .await?;
-
-                let connect = yawc::WebSocket::connect(url).fuse();
-                let timeout = executor.timer(WEBSOCKET_CONNECT_TIMEOUT).fuse();
-                futures::pin_mut!(connect, timeout);
-                let websocket = futures::select_biased! {
-                    result = connect => result.map_err(|error| anyhow!("failed to connect to Cloud WebSocket: {error}"))?,
-                    _ = timeout => return Err(anyhow!("timed out connecting to Cloud WebSocket")),
-                };
-
-                Ok(websocket::Connection::new(websocket))
-            }))
     }
 
     async fn create_llm_token(
@@ -247,10 +184,13 @@ impl CloudApiClient {
         request_builder: request::Builder,
         body: impl Into<AsyncBody>,
     ) -> Result<Response<AsyncBody>, ClientApiError> {
-        let request = {
+        let request = if matches!(self.authentication, Authentication::Session) {
+            build_request(request_builder, body, None)
+                .map_err(ClientApiError::RequestBuildFailed)?
+        } else {
             let credentials = self.credentials.read();
             let credentials = credentials.as_ref().ok_or(ClientApiError::NotSignedIn)?;
-            build_request(request_builder, body, credentials)
+            build_request(request_builder, body, Some(credentials))
                 .map_err(ClientApiError::RequestBuildFailed)?
         };
 
@@ -305,10 +245,10 @@ impl CloudApiClient {
                     .as_ref(),
             ),
             AsyncBody::default(),
-            &Credentials {
+            Some(&Credentials {
                 user_id,
                 access_token: access_token.into(),
-            },
+            }),
         )?;
 
         let mut response = self.http_client.send(request).await?;
@@ -375,13 +315,60 @@ impl CloudApiClient {
 fn build_request(
     req: request::Builder,
     body: impl Into<AsyncBody>,
-    credentials: &Credentials,
+    credentials: Option<&Credentials>,
 ) -> Result<Request<AsyncBody>> {
     Ok(req
         .header("Content-Type", "application/json")
-        .header(
-            "Authorization",
-            format!("{} {}", credentials.user_id, credentials.access_token),
-        )
+        .when_some(credentials, |request, credentials| {
+            request.header(
+                "Authorization",
+                format!("{} {}", credentials.user_id, credentials.access_token),
+            )
+        })
         .body(body.into())?)
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn build_session_authenticated_request_without_authorization_header() -> Result<()> {
+        let request = build_request(
+            Request::builder().uri("https://cloud.zed.dev/client/users/me"),
+            AsyncBody::default(),
+            None,
+        )?;
+
+        assert_eq!(
+            request
+                .headers()
+                .get("Content-Type")
+                .and_then(|value| value.to_str().ok()),
+            Some("application/json")
+        );
+        assert!(!request.headers().contains_key("Authorization"));
+        Ok(())
+    }
+
+    #[test]
+    fn build_credentials_authenticated_request_with_authorization_header() -> Result<()> {
+        let request = build_request(
+            Request::builder().uri("https://cloud.zed.dev/client/users/me"),
+            AsyncBody::default(),
+            Some(&Credentials {
+                user_id: 123,
+                access_token: "token".into(),
+            }),
+        )?;
+
+        assert_eq!(
+            request
+                .headers()
+                .get("Authorization")
+                .and_then(|value| value.to_str().ok()),
+            Some("123 token")
+        );
+        Ok(())
+    }
 }
