+++ crates/cloud_api_types/Cargo.toml
@@ -19,7 +19,6 @@ cloud_llm_client.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 strum.workspace = true
-url = { workspace = true, features = ["serde"] }
 zeta_prompt.workspace = true
 
 [dev-dependencies]
